General-purpose "Confirm" modal
Added a re-usable "yes/no" confirmation modal where the client just sets the question string and gets a callback to handle "OK". It doesn't make sense to create one modal for each confirmation, especially when it's only used in one place.

Replaced one of the existing modal as an example.