1318: Remove 'visibilitychange' listener
Per 1318, we will only be sending data when user is done with a content, so this listener is no longer needed.

I think it was also not needed when we removed the beacon usage (it's mostly to prevent data loss when beacon is used), but we didn't remove it back then to avoid changing too many things and affecting the dataset.