Latest Adnim script -- "read next" ads
(1) This replaces the bottom "sticky" with a slightly-less intrusive "read next" ad that only shows 1 ad at a time.

The location has been adjusted to reduce the overlap with the Nag, but still requires some work.

(2) The sticky was previously disabled on mobile as it was taking too much space. We'll enable the "read next" for mobile and see how it goes.

(3) The new "read next" has improved the removability, so just enable it on more pages.