Add ability to log the original err msg in `cause`
We change certain errors to make it friendlier to the user, but it won't be useful for our logs. Log both the altered and original error message.

Specifically for tus, we lost tracking which chunk got stuck. Logging it allows us to see if the user managed to progress further, or stuck at the same chunk.