Don't lazy-load ModalError
## Issue
If something wrong happens due to internet connection and we have something to convey through the modal, the message is lost because we can't fetch the lazy loaded modal.

## Change
Make ModalError part of ui.js

## Future
Any `event/desktop_errors` that failed to dispatch due to connection errors should probably be stashed and retried, so that we don't lose any valuable logs?