Optimize resolves for edited playlists
## Issue
When viewing own edited playlist (`/$/playlist/<id>`), the GUI jumped the gun and resolved entries individually because `urlsReady` is true given that edited URLs are stored in the wallet.

This is the similar problem with the private list that was fix recently, but we need to resolve the collection to get the description, but resolve the edited URLs from the wallet instead of from the claim.