Fix blank string in Error Modal on livestream-creation timeout
Livestream-creation goes through a different code path than the usual v1/v2 publishing, and there was a bug for not handling "Error vs. string" in the middle of this path.

While at it, ensured that we show the full explanation for this path instead of the generic "${method}: Your action timed out, but may have been completed.".