Fix notification popup redirects
## Issue
- Closes 1953 -- community comments not jumping to the comment directly.
- Closes 1952 -- Rewards reminder path leads to 404.

The change in d34d32ac caused the `search` parameter to be blank (everything went into `pathname`), so we always end up in the channel page's first tab.

## Change
(1) Parse the notification target and define `pathname` and `search` to feed into `NavLink::to`
(2) It seems like the `push` in `NavLink::onClick` is redundant (`onClick` and `to` is union'd?), so removed that.