TruncatedText: escape the input of `dangerouslySetInnerHTML`
Re-do of 1973 to also cover other uses of `TruncatedText`, e.g. in collection names.

On the React side of things, we typically don't need to do this because React will escape for us. The use of `dangerouslySetInnerHTML` bypassed that, so now we have to handle it.

Copying `escapeHtmlProperty` for now, as calling a `require` from `web/` feels weird. If we end up having to do more of this, then we can clone the util file over, or use `dompurify`.