(patch) show blocked mature channels on block list
## Issue
Deleted channels aren't showing up properly in the Following-Manage page

## Change
Redo 080a36c5 a bit by not making `showUserBlock` cover both abandoned and nsfw -- there is `showUnresolvedClaim` for the former already.

Technically, even this solution is wrong because `obscureNfsw` is available for that, but unfortunately this needs to fan out to `ClaimList`, `ClaimListDiscover`, etc., which is a pain to test. There is a ticket for that major refactoring.

080a36c5 just wanted to display abandoned and nsfw in the Blocklist page, so this solution still achieves that since the block list page is already passing in `showUnresolvedClaim`