Fix logging of videojs media errors
## Issue
`error()` returns a `MediaError`-like object, not `Error`-like object. Sentry hates it.

## Change
Log the individual media errors in a way that Sentry accepts.

We could now get a count for loading issues like "The media could not be loaded, either because the server or network failed or because the format is not supported.", etc.