Sentry: `whitelistUrls` doesn't work; `allowUrls` does
Documentation uses `whitelistUrls`, but saw folks mentioning in chat that it is now `allowUrls`. Tried it, and sure enough it prevented ad errors from being sent. No wonder it hasn't work in lbry.tv.

Note that it will no longer send entries in development build, because that is served by `webpack://`. Not sure whether it's wise to add that at this point.