Single source of truth for the language setting
## Issue
The previous code was trying to copy the redux's language setting into localStorage, for use in the `__( )` macro
(and perhaps also for persistence in the past -- that should now be handled by redux rehydration).

Several issues:
1. localStorage is not always available -- plenty of sentry reports on permission or quota issues. Newer browsers handle it more gracefully now, but still...
2. `doSetLanguage` is not the only place where the setting gets updated -- it can be changed from a wallet-sync as well, so the variables become out of sync.

## Change
Redux should always be the single source of truth.

We can actually expose the state to non-React code (i.e. `__( )`), and since we've already done so via `window` (not ideal, but works), we can just get the state directly from the store.  This should avoid the need to sync 2 variables.