Disable Settings Page until prefs are ready
## Reproduce
- Enter `/$/settings` directly or refresh after enter.
- Immediately change the language. 2 seconds later, the change is negated.

## Issue
Normally, changes to synced settings will be blocked with a toast, asking the user to wait for initial sync. But in Settings Page, the mechanism is bypassed since we are "locking" state changes -- but it didn't take this scenario into account.

## Change
Seems easiest to just block the page with a spinner until prefs are ready.