Don't reset language setting when fetch fails
I think resetting the language setting when the fetch fails is a bad idea, since it throws away the user's intended setting.

From sentry logs, there are occasional hiccups in the fetch, and Tom also mentioned some extensions block the fetch.

If the user changed machines or refreshes + successfully fetches, they should be getting the desired language, not the default just because of a one-time hiccup.

It should be fine for the setting to stay as-is despite the GUI not matching the selected language in this scenario. There is already a toast to indicate the failure.

If needed, we can make it clearer with a Reload/Retry button next to the language setting.