Publish:  Always show the Cancel/Remove button
The only scenario that we are hiding the button is right after `notify` is sent. For all other scenarios, including F5, we show it to allow users to clear the upload.

Since there was confusion lately, just always show it for all scenarios, but add another statement saying indicating that if the SDK already ran, they might see the item again later.