No need to clear Publish Form on success for now
## Issue
401 - publish-success modal clears Post while you are still editing.

## Notes
The original version tried to keep the form's data until Success so that we can re-upload the Post in the event of a failure.

But recent changes have introduced a `beginPublish` hook which clears the form every time you enter it. This makes the "clear on success" action no longer necessary, and fixes the annoying issue mentioned in 401.

## New Issue
The problem now becomes "data loss on page re-entry" for Post that you haven't submitted yet.

Pushing this first since it's no worse thatn status quo, and that the new issue requires a bigger change.