FileRenderInitiator: handle purchase-rent transaction
Payment prompt should have been done here in the first place (like it was with LBCs), otherwise Queue and Embeds would not work.

The logic in FileRenderInitiator has become unreadable, and the logic seems longer than necessary. Attempt to restore the older/simpler logic, and be more verbose in the variable names so that it only does 1 thing only. Recent edits seems to have overloaded things a little bit.