Publish Preview: move logic into functions. No functional change
No changes to the logic, just moving each item into functions for better management, to prepare for more items from Memberships and Paid Content.

Also re-arranged to the "props--variables--functions--effects--jsx" order.

--------

Don't like moving things around as it causes merge-conflicts and bad diffs, but no choice for this case as the component starts to get big.