Fix: no prompt to purchase
This is related to some state issues with contentRestrictedFromUser which goes from false to true, and the false value then triggered autopaly. This works for now, but that selector should be fixed.

There's still a bug with paid content, i.e. lbry://test-fee-1 - it tries to autoplay - I tried to fix it with adding canviewfile there, but then it had the opposite issue - it would go from true to false, still triggering autoplay.