(minor cleanup) Show reactions on own claims
On top of the logic error, I think I implemented incorrectly.
It should always be displayed, but can't be clicked if not paid. Should we do this instead?

Also, the current logic requires both LBC and $ to be cleared. If the claim is paywalled by both (say from Desktop or some other means), the behavior doesn't feel right.